Chore: entity class 폴더 정리
