--- src/main/java/com/nhnacademy/sensordata/entity/HumidityMaxMinMonthly.java
@@ -1,24 +1,0 @@
-package com.nhnacademy.sensordata.entity;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import org.influxdb.annotation.Column;
-import org.influxdb.annotation.Measurement;
-
-import java.time.Instant;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Measurement(name = "daily_extreme_humidity")
-public class HumidityMaxMinMonthly {
-    @Column(name = "time")
-    private Instant time;
-    @Column(name = "max_humidity")
-    private double maxHumidity;
-    @Column(name = "min_humidity")
-    private double minHumidity;
-}
