+++ src/main/java/com/nhnacademy/sensordata/entity/humidity/Humidity.java
@@ -1,19 +1,20 @@
-package com.nhnacademy.sensordata.entity;
+package com.nhnacademy.sensordata.entity.humidity;
 
 import lombok.AllArgsConstructor;
-import lombok.Data;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import org.influxdb.annotation.Column;
 import org.influxdb.annotation.Measurement;
 
 import java.time.Instant;
 
-
-@Data
+@Getter
+@Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@Measurement(name = "temperature")
-public class Temperature {
+@Measurement(name = "humidity")
+public class Humidity {
     @Column(name = "time")
     private Instant time;
     @Column(name = "device")
