+++ src/main/java/com/nhnacademy/sensordata/entity/illumination/Illumination.java
@@ -1,0 +1,26 @@
+package com.nhnacademy.sensordata.entity.illumination;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import org.influxdb.annotation.Column;
+import org.influxdb.annotation.Measurement;
+
+import java.time.Instant;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Measurement(name = "illumination")
+public class Illumination {
+    @Column(name = "time")
+    private Instant time;
+    @Column(name = "device")
+    private String device;
+    @Column(name = "place")
+    private String place;
+    @Column(name = "topic")
+    private String topic;
+    @Column(name = "value")
+    private Integer value;
+}
