+++ src/main/java/com/nhnacademy/sensordata/entity/temperature/TemperatureMaxMinWeekly.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.sensordata.entity;
+package com.nhnacademy.sensordata.entity.temperature;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -11,8 +11,8 @@ import java.time.Instant;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-@Measurement(name = "temperature_hourly")
-public class TemperatureMaxMinDaily {
+@Measurement(name = "temperature_daily")
+public class TemperatureMaxMinWeekly {
     @Column(name = "time")
     private Instant time;
     @Column(name = "max_temperature")
