+++ src/main/java/com/nhnacademy/sensordata/service/IlluminationService.java
@@ -1,6 +1,6 @@
 package com.nhnacademy.sensordata.service;
 
-import com.nhnacademy.sensordata.entity.Illumination;
+import com.nhnacademy.sensordata.entity.illumination.Illumination;
 
 public interface IlluminationService {
     Illumination getIllumination();
